[Merge] 2024-06-24 버전 적용 (#196)
* [Refactor] Emotion 추가 및 삭제 서비스 로직 오류 수정 (#195)

* [Refactor] 한줄처방 Emotion 추가 및 삭제 로직 변경

* [Refactor] 조인이 아닌 조회수 컬럼을 참조하도록 코드 변경

---------

+++ src/main/java/kr/KWGraduate/BookPharmacy/domain/onelineprescription/service/OneLineLikeEmotionService.java
@@ -11,6 +11,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class OneLineLikeEmotionService {
@@ -30,19 +32,19 @@ public class OneLineLikeEmotionService {
         // '좋아요'가 이미 존재하면 에러 반환
         if(oneLineLikeEmotionRepository.findByLoginIdAndOneLinePreId(loginId, oneLinePrescriptionId).isPresent()){
 
+        }else {
+            // 좋아요 엔티티 생성
+            OneLineLikeEmotion oneLineLikeEmotion = OneLineLikeEmotion.builder()
+                    .client(client)
+                    .oneLinePrescription(oneLinePrescription)
+                    .build();
+
+            // 한줄처방 좋아요 개수 + 1
+            oneLinePrescription.setLikeCount(oneLinePrescription.getLikeCount() + 1);
+
+            oneLineLikeEmotionRepository.save(oneLineLikeEmotion);
+            oneLinePrescriptionRepository.save(oneLinePrescription);
         }
-
-        // 좋아요 엔티티 생성
-        OneLineLikeEmotion oneLineLikeEmotion = OneLineLikeEmotion.builder()
-                .client(client)
-                .oneLinePrescription(oneLinePrescription)
-                .build();
-
-        // 한줄처방 좋아요 개수 + 1
-        oneLinePrescription.setLikeCount(oneLinePrescription.getLikeCount() + 1);
-
-        oneLineLikeEmotionRepository.save(oneLineLikeEmotion);
-        oneLinePrescriptionRepository.save(oneLinePrescription);
     }
 
     @Transactional
@@ -51,16 +53,19 @@ public class OneLineLikeEmotionService {
 
         // 예외처리는 추후에 한번에 적용하겠음
         Client client = clientRepository.findByLoginId(loginId).get();
+
         OneLinePrescription oneLinePrescription = oneLinePrescriptionRepository.findById(oneLinePrescriptionId).get();
 
-        OneLineLikeEmotion oneLineLikeEmotion = oneLineLikeEmotionRepository.findByLoginIdAndOneLinePreId(loginId, oneLinePrescriptionId).get();
+        Optional<OneLineLikeEmotion> optional = oneLineLikeEmotionRepository.findByLoginIdAndOneLinePreId(loginId, oneLinePrescriptionId);
+        if (optional.isPresent()) {
+            OneLineLikeEmotion oneLineLikeEmotion = optional.get();
 
-        // 한줄처방 좋아요 개수 - 1
-        // (한줄처방이 존재하지 않는다면 위 find함수에서 에러를 던질 것이고, 그렇게 되면 count가 감소하지 않으므로, 따로 if(count<=0)과 같은 조건처리를 해줄 필요가 없음)
-        oneLinePrescription.setLikeCount(oneLinePrescription.getLikeCount() - 1);
+            // 한줄처방 좋아요 개수 - 1
+            // (한줄처방이 존재하지 않는다면 위 find함수에서 에러를 던질 것이고, 그렇게 되면 count가 감소하지 않으므로, 따로 if(count<=0)과 같은 조건처리를 해줄 필요가 없음)
+            oneLinePrescription.setLikeCount(oneLinePrescription.getLikeCount() - 1);
 
-        oneLineLikeEmotionRepository.delete(oneLineLikeEmotion);
-        oneLinePrescriptionRepository.save(oneLinePrescription);
-
+            oneLineLikeEmotionRepository.delete(oneLineLikeEmotion);
+            oneLinePrescriptionRepository.save(oneLinePrescription);
+        }
     }
 }
